Put everything in a single module
